requirements.txt
python-telegram-bot==20.7
deep-translator==1.11.4
SpeechRecognition==3.14.4
pydub==0.25.1
ffmpeg-python==0.2.0
